fix: finish remainder of InitCustomization() modifications
In the previous commit we failed to notice the changes required in the
combine_customizations() function. Also found that you venv activation
is not idempotent wrt the prompt, so updated python.rs to perform all
pip & poetry installations in a venv-activated subshell.

Finally fixed an error in the new python-env.exp script and fine-tuned
the prompt expectations to be more specific.