fix: use bashInteractive in preference to bash for tests
Several tests have been throwing the "set: vi: invalid option name" error
on account of the fact that we were using bash rather than bashInteractive,
and several of us have spent significant time independently hunting down and
identifying this cause for the problem. This change updates the test closure
to address this problem, which should save us time in future.