docs(pkgdb): Quote package group in upgrade error (#1383)
## Proposed Changes

To make it clearer, in cases where the package doesn't belong to an
explicit package group, that `toplevel` has a special meaning as the
default group:

-
https://flox.dev/docs/reference/command-reference/manifest.toml/#package-descriptors

My first interpretation of the error was that `toplevel` was an example
rather than a reserved word.

Reformatting was applied automatically by `clang-format` to satisfy the
pre-commit hook.

## Release Notes

N/A