feat: implement refactor/redesign stages of `flox activate` (#1299)
## Proposed Changes

This PR implements the redesign of `flox activate` as discussed in
"Proposal item 3" of the following comment:
https://github.com/flox/flox/issues/1187#issuecomment-2001890585

A summary of related issues are listed in a
[comment](https://github.com/flox/flox/issues/1187#issuecomment-2046820745)
further down in that issue, some of which can be closed with this
refactor.

## Release Notes

* As before, the manifest `[hook.on-activate]` script is invoked using
the `bash` shell, but is now:
    * invoked only once with the first `flox activate` invocation
    * invoked _before_ processing the `[profile]` scripts
* Variables set in `[hook.on-activate]` will now be available for use
within `[profile]` scripts
* this is useful for starting daemons which export process/socket
information, e.g. `eval $(ssh-agent)` is probably the most commonly-used
example of this invocation idiom
* Flox environments can now be activated by invoking the
`/path/to/environment/activate` script
* this will default to using your `$SHELL` variable to detect the shell
to be invoked in the "interactive" mode, or the shell dialect to be
emitted in "in-place" activation mode
* set the `$FLOX_SHELL` environment variable to override this behaviour,
e.g. `FLOX_SHELL=bash /path/to/environment/activate`