fix: repair setting of ZDOTDIR following 1299 (#1391)
PR1299 changed the top-level "activate" path from a directory containing
scripts to a script to be invoked, and with that we changed the
"zdotdir" from an entirely separate package to instead reside within the
"activate.d" directory found in the environment itself.

This was fine for environments rendered with the new activation logic,
but obviously broke for old environment renderings which lacked this
directory.

## Proposed Changes

This patch updates the build and code to render and use a standalone
zdotdir package just as it did previously.

## Release Notes

N/A