fix(cli): make options.systems optional (#1396)
options.systems is optional for pkgdb, and if it's null, pkgdb defaults
to the current system.

The CLI currently passes an empty list, which leads to pkgdb throwing
SystemNotSupportedByLockfile.

Instead, make systems an Option in the CLI, and default to the current
system for resolution if it is None.

Drop the use of lifetimes in collect_package_groups since that would
require figuring out lifetimes for a second source of system strings
which seems like too much effort when we could just copy a few strings
instead.