fix: provide tokio to function called by spinner
The catalog client uses `reqwest` which uses `tokio::time::sleep`. When
the catalog is used in combination with our `spin_with_delay` function,
this causes a panic:
```
there is no reactor running, must be called from the context of a Tokio 1.x runtime
```

Enter the tokio runtime context in the thread spawned by
`spin_with_delay` to avoid the panic.