refactor: don't default to search result type in node hook (#1413)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
While working on the catalog init hook I discovered that we'll need a
common-denominator type for handling information retrieved by either the
catalog client or pkgdb. This introduces a new type to handle that, and
it's presented as a separate PR to keep the catalog-init PR smaller.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->