feat: use catalog client for init (#1440)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR allows using the catalog client during `flox init` via the
environment variable. Additionally, a new `just` command,
`catalog-flox`, is added to make it easy to run commands using the
catalog client as simple as `just catalog-flox <cmd>`. This PR makes the
following changes:
- Makes the init hooks async, which required removing the trait objects
in favor of dispatch via enum variant.
- Init hooks use `ProvidedPackage` as a common denominator rather than
`SearchResult` since the `/resolve` endpoint doesn't return
`SearchResult`s.
- Consolidates package fetching into function used by all init hooks.
- Introduces an `AttrPath` enum to hold attribute paths as either single
strings or as a list of attributes. This allows better compatibility
between catalog representations and pkgdb representations.
- Removes the `test_` prefix from init hook test names. This isn't
required for Rust tests as it is with `pytest`.

It's strongly encouraged to review this as individual commits.
- First commit contains a partial Node implementation plus the async
migration as this was when I discovered I'd need to do some rewriting.
- Second commit contains the remainder of the Node implementation
- Go and Python implementations are in their corresponding commits.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->