fix: don't print success before error (#1439)
Running `flox activate` twice was printing:
```
✅ You are now using the environment 'name'.
To stop using this environment, type 'exit'

ERROR: Environment 'name' is already active.
```

Remove the success message from Rust, and only print it in Bash
codepaths when the environment is not already active (and an interactive
activation is being run).

Also add a ❌ to the error message.