chore(show): Use bash package for catalog test
So that we can test what the output looks like for packages that have
multiple versions, in preparation for defaulting to `--all`.

The `hello` package doesn't any versions because it doesn't change
often but the `bash` package does.

The fixture contents were generated with the following command and then
wrapped with the additional array:

    FLOX_FEATURES_USE_CATALOG=true _FLOX_CATALOG_DUMP_RESPONSE_FILE=cli/tests/catalog_responses/bash_show.json flox show bash