feat(show): Default --all to true and deprecate
So that we expose to the user all versions of a `pkg_path` from the
catalog. There's no functional change for the non-catalog code path
because `pkgdb` only ever knows about a single version.

It looks like this defaulted to `false` because the original design
accounted for matching multiple `pkg_paths` and would later include
stabilities:

- https://github.com/flox/product/blob/main/requirements/archive/search-v1.md#design

I've left the flag, hidden it, and added a deprecation warning, in case
anyone happens to be using it currently. This can be removed in future
releases.