feat(show): Default --all to true and deprecate (#1444)
## Proposed Changes

**feat(show): Default --all to true and deprecate**

So that we expose to the user all versions of a `pkg_path` from the
catalog. There's no functional change for the non-catalog code path
because `pkgdb` only ever knows about a single version.

It looks like this defaulted to `false` because the original design
accounted for matching multiple `pkg_paths` and would later include
stabilities:

-
https://github.com/flox/product/blob/main/requirements/archive/search-v1.md#design

I've left the flag, hidden it, and added a deprecation warning, in case
anyone happens to be using it currently. This can be removed in future
releases.

**feat(show): Output one version per line**

This tidies up the output and will make it easier to copy/paste a
specific result. In turn, stop duplicating the `pkg_name`, which is
already printed on the line above with the package description.

## Release Notes

- `flox show` improvements
  - `--all` is now the default and the flag has been deprecated.
  - outputs one version per line.