build(flox-tests): export absolute default path to pkgdb (#1455)
This is 54d3ae9e4b1b988174f8710fbea5959628e5c978 but for flox-tests
instead of just flox-cli-tests.

`flox-tests` was calling `flox` with `PKGDB_PATH=pkgdb` and expecting
`flox` to find `pkgdb` in `PATH`, but since #1299 `pkgdb` is called with
a cleared `PATH`.
Unless `flox-tests` are called with `--pkgdb <abs path to pkgdb>`,
executing pkgdb will fail with a No Such File exception.