chore(show): Rename search-term to pkg-path (#1445)
## Proposed Changes

To indicate to the user that we're matching a single item rather than
searching for multiple packages with a partial string.

I've left the naming in `construct_show_params` and below because it's
specific to `pkgdb` terminology.

---

Noticed this while working on https://github.com/flox/flox/pull/1444
which it may cause merge conflicts with but thought it was worth
reviewing in its own right.

## Release Notes

N/A, probably not worth mentioning.

---------

Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>