fix: restore original ZDOTDIR after activation
This patch refactors the flox-provided `zdotdir` collection of zsh init
scripts to ensure that the original value of ZDOTDIR is restored following
activation, and amends the activation logic to be more consistent and robust
in interactive login shell case (which previously could be broken by a user's
~/.login file).

This patch also updates the README.md file to reflect the updated logic.

Finally, this patch updates the trace verbosity for zsh to selectively enable
`compinit` tracing only with verbosity levels 3 and above, whereas it had
previously been enabled at level 2. This tracing is particularly verbose,
which had significantly reduced the usefulness of the tracing.