test: enable functional catalog tests (#1458)
- Enable catalog tests instead of skipping them since serverside is
ready
- Warm up the server for search and show by running a search or show
that is allowed to fail. No warmup for install or upgrade is performed
because it hasn't been timing out in CI.
- Move the upgrade test from cli/tests to functional tests and hit the
catalog for the actual upgrade operation. The installation for an old
version is still performed using mock data. Use an actual response from
the server for that data, except manually change version to
`old_version`.

---------

Co-authored-by: billlevine <bill@floxdev.com>