feat: added tests to verify each zdotdir activation
The zdotdir activation logic is complex, and as documented in the README.md
our flox initialization script is invoked from a different RC file depending
on the way that it is invoked. This patch adds tests for each of the [3]
invocation modes, ensuring that the expected RC files and flox zsh script
are sourced in each instance.