fix: properly verify execution of zsh activation script (#1475)
## Proposed Changes

The new activate:zdotdir tests added in PR#1467 made the mistake of
exercising the ability to find an installed package in the PATH as
confirmation that the `zsh` script had run, when of course that doesn't
make sense because the PATH configuration happens in the bash activate
script that precedes it.

This patch updates the test to reuse the existing `only-once.toml`
manifest to confirm both that the `zsh` activation script is invoked to
source the `profile.zsh` section of the manifest and that it runs only
once. Each test also runs between 100-200ms faster with this change.

## Release Notes

N/A