feat: detect [hook.script] and print deprecation warning
Issue #1135 documents the request to print a deprecation warning with
the use of `[hook.script]` in a flox manifest. This patch updates the
rust `call_pkgdb()` function to pass stderr messages beginning with the
string "warning:" directly to STDERR, and similarly updates the pkgdb
`makeActivationScripts()` function to print a warning message to STDERR
when rendering the `[hook.script]` manifest section to a file.