fix: do not leak FLOX_SHELL into subsequent activations
In the following discourse posting Eric observed that when invoking
`eval "$(flox activate)"` from an `.envrc` file with `direnv` that
it was incorrectly emitting zsh-syntax output rather than detecting
that `direnv` had invoked flox via an intermediate `bash` process.
The reason for this failure was because Eric was already running in
an "outer" flox activation, and this activation had leaked the
`FLOX_SHELL` variable into his environment, which then served as an
override for direnv's use of flox.

The fix for this is to stop leaking the `FLOX_SHELL` variable into
the environment. This patch does this by unsetting the variable from
the top-level activation script immediately after having read it, and
thereafter referring to the value by way of a private read-only
`_flox_shell` variable that is not exported to the environment.

In the interim while we await the merging of this PR the workaround
for anyone using `direnv` is to explicitly set `FLOX_SHELL=bash`
from within the `.envrc` files read by `direnv`.

https://discourse.flox.dev/t/regression-with-flox-1-0-5/899