fix: help and docs inaccuracies (#1487)
Drop `semver` in the manifest man page since it isn't supported. Trying
to use it results in:
```
error parsing manifest descriptor: encountered unrecognized field 'semver' while parsing manifest descriptor
```

Merge the description of `<package>` from `flox install --help` with
that in the man page. The two should be consistent, and the current help
description uses `relative path` which is no longer a term that should
be user facing.

Co-authored-by: Bryan Honof <bryan@floxdev.com>