fix: source ~/.tcshrc only once per activation, ++
Matthew spotted that we were sourcing ~/.tcshrc twice, so created the
activate:validate_hook_and_dotfile_sourcing test to confirm that we
are sourcing all the expected files in exactly the right order, and
only once, for all shells. Then used this test to fix the tcsh order
of sourcing the various scripts and hooks.

Took this opportunity to make consistent the STDERR messages coming
out of the various dotfiles in the fish and tcsh cases, and also
addressed problems with tests interfering with each other as they
stomped over $HOME/.{zshrc,bashrc,...}. My approach there was to
create an "*.extra" convention by which any test could _add_ to what
the default dotfiles were doing, and then remove the "*.extra" file
at the conclusion of the test.