fix: remove interactive (-i) flag from activate bash test
In creating the activate:validate_hook_and_dotfile_sourcing test in the
activate.bats suite I copied the shell invocation arguments from a previous
test, but as it happens invoking bash with `-i` causes the following error
messages to be emitted when running in CI:

  bash: cannot set terminal process group (28939): Inappropriate ioctl for device
  bash: no job control in this shell

Removing -i from the bash invocation, as well as for the other shells.

The zsh example is slightly different because it's designed not to complain
about such things when invoked with `-i`, so I left it in place for consistency
and so that the scripts and order being sourced did not change.