fix(activate): Convert tests from expect
We've had another flakey test:

- https://github.com/flox/flox/actions/runs/9377878053/job/25820156213#step:6:84

I don't think that `expect` is necessarily at fault here but the
timeouts and lack of feedback make it very hard to understand what's
happened. We could enable `exp_internal 1` to get more feedback but it's
very noisy, especially combined with `activate -v`.

Converting the assertions to a straight shell script that runs within
the activation is slightly faster but more importantly easier to reason
about and gives us clearer feedback. We took the same approach for
Python tests in the past.

I believe that these tests are functionally still the same. I've left
`hello.exp` in place because it's still used by one other test that has
a more specific assertion about the binary path.