fix: invoke `tcsh -m` to source flox-provided $HOME/.tcshrc (#1536)
## Proposed Changes

By default tcsh silently refuses to source a .tcshrc file that the
effective user does not own, but this behaviour can be overridden with
the `-m` flag as described in tcsh(1):

-m The shell loads ˜/.tcshrc even if it does not belong to the effective
user.

This patch adds this `-m` flag to the two tcsh invocations for which we
are setting `$HOME`, and removes the `-v` flag that was accidentally
left behind with a previous round of debugging.

## Release Notes

N/A

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>