test(environment-managed): Copy cases for catalog (#1543)
## Proposed Changes

**test(environment-managed): Copy cases for catalog**

Duplicate the tests that would touch the catalog to run in parallel with
pkgdb test cases. They are modified to pass in a subsequent commit to
make it easier to compare against the originals.

**test(environment-managed): Add catalog mock resps**

For all places where we install a package and expect to resolve from the
catalog server. These are specified inline so that all other commands
default to using the empty response and will blow up if we don't provide
the correct mocks.

**test(environment-managed): Fix manifest version**

Add a `version` when constructing a catalog compatible manifest to fix
this error:

    ❌ ERROR: unrecognized manifest field: 'version'.

**test(environment-managed): Fix manifest pkg-path**

Add a `pkg-path` when constructing a catalog compatible manifest to fix
this error:

    ❌ ERROR: Failed to parse manifest: missing field `pkg-path`
    in `install.hello`

---

I don't think it's worth attempting to de-duplicate the installs at the
moment while we have parallel pkgdb and catalog tests. It'll be easier
to do afterwards.

## Release Notes

N/A