test(environment-managed): Use non-inter activate
This test was broken by asserting on a different prompt for
`activate.bats` in a preceding commit.

Rather than adapt the assertion for both prompts or duplicate the tests,
I think it's safe to convert this to a non-interactive use of `flox
activate`, because we have good coverage of all the interactive cases in
`activate.bats` and they shouldn't be affected by other types of
environments.

The comment about `USER` and `REAL_USER` appears to have been copied
from `activate.bats` and doesn't seem strictly relevant to the behaviour
we're testing here.

The regex needed tweaking because the extra `\..+` was being greedy and
incorrectly matching across multiple lines of output from `expect`.