test(environment-remote): Use non-inter activate
This test was susceptible to the same problem as `activate.bats` where
we were matching on a spinner that would sometimes not be rendered if
the execution was fast enough.

Rather than adapt the assertion for both prompts or duplicate the tests,
I think it's safe to convert this to a non-interactive use of `flox
activate`, because we have good coverage of all the interactive cases in
`activate.bats` and they shouldn't be affected by other types of
environments.

The comment about `USER` and `REAL_USER` appears to have been copied
from `activate.bats` and doesn't seem strictly relevant to the behaviour
we're testing here.