test: Fix interactive prompt assertion and convert others to non-interactive (#1547)
## Proposed Changes

**test(activate): Assert on interactive prompt**

This both addresses the comment about us not asserting on the prompt and
fixes this intermittent test failure:

-
https://github.com/flox/flox/actions/runs/9402290739/job/25896156760#step:6:235

Which is thought, now that we get verbose output from expect, to happen
because the activation happened too quickly for the spinner to render:

# expect: does "flox [project-11] build-intel% \u001b[?2004h" (spawn_id
exp6) match glob pattern "Preparing environment"? no
    # "*\n"? no
    # "*\r"? no
    # expect: timed out

This will break the use of `interactive-hello.exp` by
`environment-managed.bats`, because it generates a different prompt,
which I'll fix in a subsequent commit.

**test(environment-managed): Use non-inter activate**

This test was broken by asserting on a different prompt for
`activate.bats` in a preceding commit.

Rather than adapt the assertion for both prompts or duplicate the tests,
I think it's safe to convert this to a non-interactive use of `flox
activate`,
because we have good coverage of all the interactive cases in
`activate.bats` and they shouldn't be affected by other types of
environments.

The comment about `USER` and `REAL_USER` appears to have been copied
from `activate.bats` and doesn't seem strictly relevant to the behaviour
we're testing here.

The regex needed tweaking because the extra `\..+` was being greedy and
incorrectly matching across multiple lines of output from `expect`.

**test(environment-remote): Use non-inter activate**

This test was susceptible to the same problem as `activate.bats` where
we were matching on a spinner that would sometimes not be rendered if
the execution was fast enough.

Rather than adapt the assertion for both prompts or duplicate the tests,
I think it's safe to convert this to a non-interactive use of `flox
activate`, because we have good coverage of all the interactive cases in
`activate.bats` and they shouldn't be affected by other types of
environments.

The comment about `USER` and `REAL_USER` appears to have been copied
from `activate.bats` and doesn't seem strictly relevant to the behaviour
we're testing here.

## Release Notes

N/A