test(ld-floxlib): Adapt test for catalog
The project setup had to be split up to run pkgdg or catalog specific
commands from each test case.

The manifest needed to be forked to include a version and the specific
versions of `glibc` and `nix` that are required by the test. I couldn't
resolve the original version of `nix` from the catalog but it seems that
the specific version is only need to assert on the output, which has
been updated.

I also had to generate the mock response without `generate-reponses.sh`
because we need to provide a full manifest (rather than the install
commands) and the use of `glibc` restricts it to Linux. This should be
easier when we have the `mk_data` script though.