fix: Handle broken symlinks in run/ directory
In the case where the user has uninstalled and reinstalled Flox, or if
the environment is somehow garbage collected, the symlink to the
environment in .flox/env/run will be broken. Previously, we got an error
in that case. We now treat it as invalidated the same as if the
environment needed a rebuild, and just try to build the environment
again.