test(ld-floxlib): Copy and adapt tests for catalog (#1551)
## Proposed Changes

**test(ld-floxlib): Copy tests for catalog**

Duplicate the test that would touch the catalog to run in parallel with
pkgdb test case. They are modified to pass in a subsequent commit, to
make it easier to compare against the original.

**test(ld-floxlib): Adapt test for catalog**

The project setup had to be split up to run pkgdg or catalog specific
commands from each test case.

The manifest needed to be forked to include a version and the specific
versions of `glibc` and `nix` that are required by the test. I couldn't
resolve the original version of `nix` from the catalog but it seems that
the specific version is only need to assert on the output, which has
been updated.

I also had to generate the mock response without `generate-reponses.sh`
because we need to provide a full manifest (rather than the install
commands) and the use of `glibc` restricts it to Linux. This should be
easier when we have the `mk_data` script though.

**test(ld-floxlib): Uncomment rm in project-teardown**

It looks like the comment may have been left for testing and is no
longer required.

## Release Notes

N/A