fix: require descriptor system enabled in options or default (#1558)
An error should be raised if a manifest descriptor tries to lock a
system that is not specified in `options.systems` i.e. iff
`install.*.systems ⊈ (options.systems ? {default systems..})`.

Fixes a logic regression wrt to pkgdb based locking which required this
invariant but implemented it in pkgbd causing it to be overlooked while
implementing the rust equivalent.