feat(search): Strip version specifiers for catalog
The catalog server won't support searching by version for now. You can
use `flox show` to find the versions for a given package:

- https://github.com/flox/catalog-server/issues/60

The server has been stripping the version specifier server-side.
Implementing this client-side will allow us to remove that, which makes
it clearer, and easier to support versions as a separate parameter in
the future:

- https://github.com/flox/catalog-server/issues/98

The note about version specifiers has been removed from the `--help`
message because we're not able to switch it based on the catalog feature
flag but the existing behaviour will continue to work for pkgdb.

It feels right to implement the logic in the `catalog`, but not as a
method on the client because it doesn't need any state, and then
generate the actual warning message in `search` where we're closer to
the user interface.

I've left one integration test to verify that the warning is generated.
The previous search term variations are now covered by a unit test. We
don't have a way to assert on the search term that is actually sent to
the catalog server but that feels OK.