feat(catalog): support unfree, broken, licenses (#1572)
Unfree, broken, and licenses are already copied from the manifest into
catalog resolution requests. But pkgdb warnings and errors still needed
to be ported to catalog code paths.

- When `options.allow` are changed with a `flox edit`, an error should
  be thrown if already installed packages are no longer allowed
- When unfree or broken packages are installed, a warning should be
  printed to the user.

Note that with pkgdb, if a package failed to install because it was not
allowed, a helpful error was printed suggesting a modification to
`options.allow`. This is no longer possible since resolution failures do
not currently include enough information to determine if the failure was
due to an `options.allow` setting. The corresponding copied but skipped
integration tests were dropped instead of re-enabling.

`fake_package` was moved to a `test_helpers` module, and
`resolved_pkg_group_with_dummy_package` was moved from the `flox` crate
to the `flox-rust-sdk` crate so these helpers could be used in unit
tests.