refactor: parse package descriptor in rust (#1562)
Implements descriptor argument parsing in flox-rust-sdk rather than
calling pkgdb.

The new parser is strictly weaker parsing only an attribute path and
version, in line with the requirements for catalog based installs and
the utilized capabilities of pkgdb.
In particular support for declaring inputs and systems in a single cli
argument are dropped.