feat(search): Strip version specifiers for catalog (#1560)
## Proposed Changes

**feat(search): Strip version specifiers for catalog**

The catalog server won't support searching by version for now. You can
use `flox show` to find the versions for a given package:

- flox/catalog-server#60

The server has been stripping the version specifier server-side.
Implementing this client-side will allow us to remove that, which makes
it clearer, and easier to support versions as a separate parameter in
the future:

- flox/catalog-server#98

The note about version specifiers has been removed from the `--help`
message because we're not able to switch it based on the catalog feature
flag but the existing behaviour will continue to work for pkgdb.

It feels right to implement the logic in the `catalog`, but not as a
method on the client because it doesn't need any state, and then
generate the actual warning message in `search` where we're closer to
the user interface.

I've left one integration test to verify that the warning is generated.
The previous search term variations are now covered by a unit test. We
don't have a way to assert on the search term that is actually sent to
the catalog server but that feels OK.

**docs(search): Remove pkgdb and update for catalog**

These details aren't applicable when the catalog feature flag is used
and can be removed when we published the catalog version of these docs.

**docs(search): Catalog matches pkg-path or desc**

Bill described the catalog service doing the following:

> We do a fuzzy match against the last tuple of the `attr_path` the
> `description`, and the `name`.  Currently the `name` is transitively
> the same as the `attr_path` though, and I don't think I would document
> that as part of the search at this time.

## Release Notes

N/A until catalog release.