fix: ensure _flox_activate_tracelevel not unset by "inner" activations
Running `eval "$(flox activate)"` from within a dotfile has the effect of
sourcing the shell-specific flox init script which finishes by cleaning
the `_flox_activate_tracelevel` variable from the environment. This is
fine in an interactive setting, but when this is happening within an
"inner" in-place activation launched from an "outer" interactive one,
then this causes problems for the latter.

The fix is to expressly guard against the unsetting of this variable
as can happen when sourcing any "dotfile" that may contain an invocation
of `eval "$(flox activate)"` or equivalent.

The bug existed for each of the bash, fish, and tcsh shells, and this
patch includes fixes for all three. This bug did not exist for zsh on
account of the elaborate "zdotdir" system of zsh files which works
differently.