feat: Provide platform specific update instructions when available (#1557)
## Proposed Changes

This change provides a way to have platform specific update instructions
(for flox, not flox environments).

How to upgrade should be in the domain of the installers, since simple
platform detection is not going to work in all cases. An example where
this won't work is macOS, where we can not know how flox was installed
except if we record this at the installation time.

This PR looks up for `update-instructions.txt` file (usually in
`/usr/local/share/flox/files` folder) if there are any custom update
instructions and prints them for the user. In case
`update-instructions.txt` file can not be found, we default to current
update message.


## Release Notes

fixes #1304